blind_coding

+++ app/Http/Controllers/BulkSMSController.php
@@ -46,7 +46,12 @@ class BulkSMSController extends Controller
     {
         // return the contact group page
         $data['user'] = $user = Auth::user();
+        
         $data['contact'] = $contact = ContactGroup::find($id);
+        $user = Auth::user();
+        if($user->id !== $contact->user_id) {
+            return redirect()->back()->with('message','Access Denied');
+        }
         if ($contact->type == 'import') {
             $data['contacts'] = implode(',', unserialize($contact->contacts));
         } else {
@@ -142,6 +147,10 @@ class BulkSMSController extends Controller
     {
         // function to delete created contact group.
         $contact = ContactGroup::find($id);
+        $user = Auth::user();
+        if($user->id !== $contact->user_id) {
+            return redirect()->back()->with('message','Access Denied');
+        }
         $contact->delete();
         return redirect()->back()->with("message", "Contact group deleted succcesfully!");
     }
@@ -422,6 +431,10 @@ class BulkSMSController extends Controller
     public function resendSMS($id)
     {
         $tranx = Transaction::find($id);
+        $user = Auth::user();
+        if($user->id !== $tranx->user_id) {
+            return redirect()->back()->with('message','Access Denied');
+        }
         // dd($tranx);
         $the_work = $this->sendSMS($tranx->sender, $tranx->recipient, $tranx->message, $tranx->amount, $tranx->message_type);
